pyproject.toml
[project]
name = "driverless-agi"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["openai", "pyyaml", "python-dotenv", "nicegui>=1.4.0", "markdown>=3.5", "ddgs>=2.0", "httpx>=0.27", "beautifulsoup4>=4.12", "matplotlib>=3.7"]

[project.scripts]
dagi = "main:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends.legacy:build"
